external-api -> mvc-api 모듈 네이밍 변경
