--- external-api/Dockerfile
@@ -1,7 +1,0 @@
-FROM openjdk:21-jdk-slim
-
-WORKDIR /app
-
-COPY build/libs/*.jar external-api.jar
-
-ENTRYPOINT ["java", "-jar", "external-api.jar"]
