--- external-api/src/main/kotlin/com/jinuk/toy/externalapi/ExternalApplication.kt
@@ -1,17 +1,0 @@
-package com.jinuk.toy.externalapi
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration
-import org.springframework.boot.runApplication
-
-@SpringBootApplication(
-    scanBasePackages = ["com.jinuk.toy"],
-    exclude = [
-        DataSourceAutoConfiguration::class,
-    ],
-)
-class ExternalApplication
-
-fun main(args: Array<String>) {
-    runApplication<ExternalApplication>(*args)
-}
