--- external-api/src/main/kotlin/com/jinuk/toy/externalapi/global/ExternalAPIController.kt
@@ -1,8 +1,0 @@
-package com.jinuk.toy.externalapi.global
-
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-@RequestMapping("/api")
-annotation class ExternalAPIController
