--- external-api/src/main/kotlin/com/jinuk/toy/externalapi/global/exception/ErrorResponse.kt
@@ -1,9 +1,0 @@
-package com.jinuk.toy.externalapi.global.exception
-
-data class ErrorResponse(
-    val message: String,
-) {
-    companion object {
-        fun from(message: String?) = ErrorResponse(message ?: "잘못된 요청입니다.")
-    }
-}
