--- external-api/src/main/kotlin/com/jinuk/toy/externalapi/view/post/response/PostDetailResponse.kt
@@ -1,35 +1,0 @@
-package com.jinuk.toy.externalapi.view.post.response
-
-import com.jinuk.toy.applicaiton.post.query.result.PostDetailResult
-import com.jinuk.toy.domain.post.value.PostTitle
-import com.jinuk.toy.domain.user.value.Username
-import java.time.LocalDateTime
-
-data class PostDetailResponse(
-    val id: Long?,
-    val userId: Long,
-    val username: Username,
-    val title: PostTitle,
-    val content: String,
-    val isViewerLike: Boolean,
-    val likeCount: Int,
-    val commentCount: Int,
-    val createdAt: LocalDateTime,
-    val updatedAt: LocalDateTime,
-)
-
-internal fun PostDetailResult.toResponse() =
-    with(this) {
-        PostDetailResponse(
-            id = id,
-            userId = userId,
-            username = username,
-            title = title,
-            content = content,
-            isViewerLike = isViewerLike,
-            likeCount = likeCount,
-            commentCount = commentCount,
-            createdAt = createdAt,
-            updatedAt = updatedAt,
-        )
-    }
