+++ mvc-api/src/main/kotlin/com/jinuk/toy/mvcapi/MvcAPIApplication.kt
@@ -1,0 +1,17 @@
+package com.jinuk.toy.mvcapi
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration
+import org.springframework.boot.runApplication
+
+@SpringBootApplication(
+    scanBasePackages = ["com.jinuk.toy"],
+    exclude = [
+        DataSourceAutoConfiguration::class,
+    ],
+)
+class MvcAPIApplication
+
+fun main(args: Array<String>) {
+    runApplication<MvcAPIApplication>(*args)
+}
