+++ mvc-api/src/main/kotlin/com/jinuk/toy/mvcapi/global/MvcAPIController.kt
@@ -1,0 +1,8 @@
+package com.jinuk.toy.mvcapi.global
+
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+@RequestMapping("/api")
+annotation class MvcAPIController
