+++ mvc-api/src/main/kotlin/com/jinuk/toy/mvcapi/global/exception/ErrorResponse.kt
@@ -1,0 +1,9 @@
+package com.jinuk.toy.mvcapi.global.exception
+
+data class ErrorResponse(
+    val message: String,
+) {
+    companion object {
+        fun from(message: String?) = ErrorResponse(message ?: "잘못된 요청입니다.")
+    }
+}
