+++ mvc-api/src/main/kotlin/com/jinuk/toy/mvcapi/global/security/SecurityConfig.kt
@@ -1,0 +1,44 @@
+package com.jinuk.toy.mvcapi.global.security
+
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity
+import org.springframework.security.config.annotation.web.builders.HttpSecurity
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
+import org.springframework.security.config.http.SessionCreationPolicy
+import org.springframework.security.web.SecurityFilterChain
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
+import org.springframework.web.cors.CorsConfigurationSource
+import org.springframework.web.cors.CorsUtils
+
+@Configuration
+@EnableMethodSecurity(securedEnabled = true)
+@EnableWebSecurity
+class SecurityConfig(
+    private val userAuthenticationFilter: UserAuthenticationFilter,
+) {
+    @Bean
+    fun filterChain(
+        http: HttpSecurity,
+        corsConfigurationSource: CorsConfigurationSource,
+    ): SecurityFilterChain {
+        http
+            .headers { headerConfig -> headerConfig.frameOptions { it.disable() } }
+            .csrf { it.disable() }
+            .cors { it.disable() }
+            .httpBasic { it.disable() }
+            .formLogin { it.disable() }
+            .rememberMe { it.disable() }
+            .sessionManagement { it.sessionCreationPolicy(SessionCreationPolicy.STATELESS) }
+            .authorizeHttpRequests {
+                it.requestMatchers(CorsUtils::isPreFlightRequest).permitAll()
+                    .anyRequest().permitAll()
+            }
+            .addFilterBefore(
+                userAuthenticationFilter,
+                UsernamePasswordAuthenticationFilter::class.java,
+            )
+
+        return http.build()
+    }
+}
