+++ mvc-api/src/main/kotlin/com/jinuk/toy/mvcapi/view/auth/exception/AuthAPIExceptionHandler.kt
@@ -1,0 +1,23 @@
+package com.jinuk.toy.mvcapi.view.auth.exception
+
+import com.jinuk.toy.mvcapi.global.exception.ErrorResponse
+import com.jinuk.toy.mvcapi.view.auth.AuthAPI
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.ExceptionHandler
+import org.springframework.web.bind.annotation.RestControllerAdvice
+
+@RestControllerAdvice(assignableTypes = [AuthAPI::class])
+class AuthAPIExceptionHandler {
+    @ExceptionHandler(IllegalArgumentException::class)
+    fun handleIllegalArgumentException(e: IllegalArgumentException) =
+        ResponseEntity
+            .status(HttpStatus.BAD_REQUEST)
+            .body(ErrorResponse.from(e.message))
+
+    @ExceptionHandler(NoSuchElementException::class)
+    fun handleIllegalArgumentException(e: NoSuchElementException) =
+        ResponseEntity
+            .status(HttpStatus.NOT_FOUND)
+            .body(ErrorResponse.from(e.message))
+}
