+++ mvc-api/src/main/kotlin/com/jinuk/toy/mvcapi/view/auth/request/AuthCredentialsRequest.kt
@@ -1,0 +1,11 @@
+package com.jinuk.toy.mvcapi.view.auth.request
+
+import com.jinuk.toy.applicaiton.auth.command.usecase.SignupCommand
+import com.jinuk.toy.applicaiton.auth.query.usecase.LoginQuery
+import com.jinuk.toy.domain.user.value.Username
+
+data class AuthCredentialsRequest(val username: Username, val password: String)
+
+internal fun AuthCredentialsRequest.toQuery() = LoginQuery(username, password)
+
+internal fun AuthCredentialsRequest.toCommand() = SignupCommand(username, password)
